Add pseudo random port scanning
This is based on the Linear Congruential Generator algorithm which
guarantees we will iterate through an entire range in a pseudo random
manner.

This implementation should be bound in memory by the batch_size
specified by the user, which will be in average 1_000.

+++ src/scanner.rs
@@ -20,7 +20,7 @@ pub struct Scanner {
     host: IpAddr,
     start: u16,
     end: u16,
-    batch_size: u32,
+    batch_size: u16,
     timeout: Duration,
     quiet: bool,
 }
@@ -30,7 +30,7 @@ impl Scanner {
         host: IpAddr,
         start: u16,
         end: u16,
-        batch_size: u32,
+        batch_size: u16,
         timeout: Duration,
         quiet: bool,
     ) -> Self {
